requirements.txt
black==20.8b1
cached-property==1.5.2
camera-calibration-parsers==1.12.0
-e git+https://gitlab.mrt.uni-karlsruhe.de/zanker/dataset-utilities.git@f115fee055198b35cc15cab004b759ae88141333#egg=dataset_utilities
descartes==1.1.0
Fiona==1.8.18
flake8==3.8.4
future==0.18.2
h5py==3.1.0
importlib-metadata==3.1.0
matplotlib==3.3.3
numpy==1.18.4
opencv-python==4.4.0.46
Pillow==8.0.1
pyquaternion==0.9.9
PyYAML==5.3.1
scipy==1.5.4
Shapely==1.7.1
threadpoolctl==2.1.0
tqdm==4.53.0
